requirements.txt
#Github.com-Vasusen-code

https://github.com/vasusen-code/Telethon/archive/refs/tags/v1.24.0.zip
https://github.com/vasusen-code/ethon/archive/refs/tags/v0.1.5.zip
cryptg
tgcrypto
pyrogram
python-decouple
logzero
moviepy